Split SearchBar and SelectedCities into separate components
